[ refactor ] 사용자 정보 관련 DTO 클래스 네이밍 및 구조 변경 ( #17 )
